+++ lib/cluster/generate.go
@@ -1,4 +1,0 @@
-package cluster
-
-//go:generate ./paths_generate.lua
-//go:generate goimports -w paths.go